requirements.txt
Appium-Python-Client==0.48
asn1crypto==1.2.0
blinker==1.4
Brotli==1.0.7
certifi==2019.11.28
cffi==1.13.2
Click==7.0
cryptography==2.4.2
Flask==1.1.1
h11==0.9.0
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kaitaistruct==0.8
ldap3==2.6.1
MarkupSafe==1.1.1
mitmproxy==5.0.0
passlib==1.7.2
protobuf==3.10.0
publicsuffix2==2.20190812
pyasn1==0.4.8
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyperclip==1.7.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
selenium==3.141.0
six==1.13.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.12
tornado==6.0.3
urllib3==1.25.7
urwid==2.0.1
Werkzeug==0.16.0
wsproto==0.14.1
zstandard==0.12.0
